Feat : HomePage 로직 분리
